+++ app/mpu6050/mpu6050.hpp
@@ -26,25 +26,25 @@ namespace MPU6050 {
         ~MPU6050() noexcept;
 
         /* celsius */
-        std::optional<std::float32_t> get_temperature_celsius() const noexcept;
+        std::optional<std::float64_t> get_temperature_celsius() const noexcept;
 
         /* meters per square second */
-        std::optional<Vec3D<std::float32_t>> get_acceleration_scaled() const noexcept;
-        std::optional<std::float32_t> get_acceleration_x_scaled() const noexcept;
-        std::optional<std::float32_t> get_acceleration_y_scaled() const noexcept;
-        std::optional<std::float32_t> get_acceleration_z_scaled() const noexcept;
+        std::optional<Vec3D<std::float64_t>> get_acceleration_scaled() const noexcept;
+        std::optional<std::float64_t> get_acceleration_x_scaled() const noexcept;
+        std::optional<std::float64_t> get_acceleration_y_scaled() const noexcept;
+        std::optional<std::float64_t> get_acceleration_z_scaled() const noexcept;
 
         /* radians */
-        std::optional<Vec3D<std::float32_t>> get_rotation_scaled() const noexcept;
-        std::optional<std::float32_t> get_rotation_x_scaled() const noexcept;
-        std::optional<std::float32_t> get_rotation_y_scaled() const noexcept;
-        std::optional<std::float32_t> get_rotation_z_scaled() const noexcept;
+        std::optional<Vec3D<std::float64_t>> get_rotation_scaled() const noexcept;
+        std::optional<std::float64_t> get_rotation_x_scaled() const noexcept;
+        std::optional<std::float64_t> get_rotation_y_scaled() const noexcept;
+        std::optional<std::float64_t> get_rotation_z_scaled() const noexcept;
 
         /* degrees */
-        std::optional<Vec3D<std::float32_t>> get_roll_pitch_yaw() const noexcept;
-        std::optional<std::float32_t> get_roll() const noexcept;
-        std::optional<std::float32_t> get_pitch() const noexcept;
-        std::optional<std::float32_t> get_yaw() const noexcept;
+        std::optional<Vec3D<std::float64_t>> get_roll_pitch_yaw() const noexcept;
+        std::optional<std::float64_t> get_roll() const noexcept;
+        std::optional<std::float64_t> get_pitch() const noexcept;
+        std::optional<std::float64_t> get_yaw() const noexcept;
 
         std::uint8_t read_byte(std::uint8_t const reg_address) const noexcept;
 
@@ -239,8 +239,8 @@ namespace MPU6050 {
 
         I2CDevice i2c_device_{};
 
-        std::float32_t gyro_scale_{};
-        std::float32_t accel_scale_{};
+        std::float64_t gyro_scale_{};
+        std::float64_t accel_scale_{};
     };
 
 }; // namespace MPU6050
